requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
distlib==0.3.9
Django==5.0.6
django-ckeditor==6.7.3
django-js-asset==2.2.0
django-livereload-server==0.5.1
filelock==3.16.1
idna==3.6
pillow==11.3.0
pip==24.2
platformdirs==4.3.6
python-decouple==3.8
pytz==2024.1
requests==2.31.0
sqlparse==0.5.0
swapper==1.3.0
tabulate==0.9.0
tornado==6.4
tzdata==2024.1
urllib3==2.2.1
uv==0.1.17
virtualenv==20.27.0
